Changed a lot

+++ src/main/java/com/eEducation/ftn/model/CollegeDirection.java
@@ -39,7 +39,7 @@ public class CollegeDirection implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "numOfYears")
@@ -50,15 +50,15 @@ public class CollegeDirection implements Serializable {
     public CollegeDirection() {
     }
 
-    public CollegeDirection(Integer id) {
+    public CollegeDirection(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
