+++ src/main/java/com/eEducation/ftn/model/Course.java
@@ -41,7 +41,7 @@ public class Course implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "description")
@@ -69,15 +69,15 @@ public class Course implements Serializable {
     public Course() {
     }
 
-    public Course(Integer id) {
+    public Course(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
