+++ src/main/java/com/eEducation/ftn/model/CourseFile.java
@@ -43,7 +43,7 @@ public class CourseFile implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "documentName")
     private String documentName;
     @Column(name = "documentType")
@@ -64,15 +64,15 @@ public class CourseFile implements Serializable {
     public CourseFile() {
     }
 
-    public CourseFile(Integer id) {
+    public CourseFile(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
