+++ src/main/java/com/eEducation/ftn/model/CourseLesson.java
@@ -41,7 +41,7 @@ public class CourseLesson implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "description")
@@ -55,15 +55,15 @@ public class CourseLesson implements Serializable {
     public CourseLesson() {
     }
 
-    public CourseLesson(Integer id) {
+    public CourseLesson(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
