+++ src/main/java/com/eEducation/ftn/model/ExamTerm.java
@@ -40,7 +40,7 @@ public class ExamTerm implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @JoinColumn(name = "courseId", referencedColumnName = "id")
     @ManyToOne
     private Course course;
@@ -56,15 +56,15 @@ public class ExamTerm implements Serializable {
     public ExamTerm() {
     }
 
-    public ExamTerm(Integer id) {
+    public ExamTerm(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
