+++ src/main/java/com/eEducation/ftn/model/Grade.java
@@ -38,7 +38,7 @@ public class Grade implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
     @Column(name = "points")
     private Float points;
@@ -54,15 +54,15 @@ public class Grade implements Serializable {
     public Grade() {
     }
 
-    public Grade(Integer id) {
+    public Grade(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
