+++ src/main/java/com/eEducation/ftn/model/Notification.java
@@ -41,7 +41,7 @@ public class Notification implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "message")
     private String message;
     @Column(name = "nDate")
@@ -62,15 +62,15 @@ public class Notification implements Serializable {
     public Notification() {
     }
 
-    public Notification(Integer id) {
+    public Notification(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
