+++ src/main/java/com/eEducation/ftn/model/Payment.java
@@ -43,7 +43,7 @@ public class Payment implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @JoinColumn(name = "studentId", referencedColumnName = "id")
     @ManyToOne
     private Student student;
@@ -61,15 +61,15 @@ public class Payment implements Serializable {
     public Payment() {
     }
 
-    public Payment(Integer id) {
+    public Payment(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
