+++ src/main/java/com/eEducation/ftn/model/Rank.java
@@ -38,7 +38,7 @@ public class Rank implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @OneToMany(mappedBy = "rank")
@@ -47,15 +47,15 @@ public class Rank implements Serializable {
     public Rank() {
     }
 
-    public Rank(Integer id) {
+    public Rank(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
