+++ src/main/java/com/eEducation/ftn/model/StudentAttendsCourse.java
@@ -36,7 +36,7 @@ public class StudentAttendsCourse implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @JoinColumn(name = "studentId", referencedColumnName = "id")
     @ManyToOne
     private Student student;
@@ -47,15 +47,15 @@ public class StudentAttendsCourse implements Serializable {
     public StudentAttendsCourse() {
     }
 
-    public StudentAttendsCourse(Integer id) {
+    public StudentAttendsCourse(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
