+++ src/main/java/com/eEducation/ftn/model/StudentDocument.java
@@ -40,7 +40,7 @@ public class StudentDocument implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "documentName")
     private String documentName;
     @Column(name = "documentType")
@@ -59,15 +59,15 @@ public class StudentDocument implements Serializable {
     public StudentDocument() {
     }
 
-    public StudentDocument(Integer id) {
+    public StudentDocument(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
