+++ src/main/java/com/eEducation/ftn/model/StudentExamEntry.java
@@ -36,7 +36,7 @@ public class StudentExamEntry implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @JoinColumn(name = "studentId", referencedColumnName = "id")
     @ManyToOne
     private Student student;
@@ -50,15 +50,15 @@ public class StudentExamEntry implements Serializable {
     public StudentExamEntry() {
     }
 
-    public StudentExamEntry(Integer id) {
+    public StudentExamEntry(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
