+++ src/main/java/com/eEducation/ftn/model/Teacher.java
@@ -43,7 +43,7 @@ public class Teacher implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "firstname")
     private String firstname;
     @Column(name = "lastname")
@@ -63,15 +63,15 @@ public class Teacher implements Serializable {
     public Teacher() {
     }
 
-    public Teacher(Integer id) {
+    public Teacher(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
