+++ src/main/java/com/eEducation/ftn/service/PaymentService.java
@@ -15,7 +15,7 @@ public class PaymentService {
 	@Autowired
 	PaymentRepository paymentRepository;
 	
-	public Payment findOne(Integer id) {
+	public Payment findOne(Long id) {
 		return paymentRepository.findOne(id);
 	}
 
@@ -31,7 +31,7 @@ public class PaymentService {
 		return paymentRepository.save(payment);
 	}
 
-	public void remove(Integer id) {
+	public void remove(Long id) {
 		paymentRepository.delete(id);
 	}
 }
