+++ src/main/java/com/eEducation/ftn/service/StudentDocumentService.java
@@ -15,7 +15,7 @@ public class StudentDocumentService {
 	@Autowired
 	StudentDocumentRepository sdRepository;
 	
-	public StudentDocument findOne(Integer id) {
+	public StudentDocument findOne(Long id) {
 		return sdRepository.findOne(id);
 	}
 
@@ -31,7 +31,7 @@ public class StudentDocumentService {
 		return sdRepository.save(doc);
 	}
 
-	public void remove(Integer id) {
+	public void remove(Long id) {
 		sdRepository.delete(id);
 	}
 }
